+++ stats/src/MatchReader.ts
@@ -1,19 +1,8 @@
-import { CsvFileReader } from './CsvFileReader';
-import { dateStringToDate } from './utils';
-import { MatchResult } from './MatchResult';
+interface DataReader {
+  read(): void;
+  data: string[][];
+}
 
-type MatchData = [Date, string, string, number, number, MatchResult, string];
-
-export class MatchReader extends CsvFileReader<MatchData> {
-  mapRow(row: string[]): MatchData {
-    return [
-      dateStringToDate(row[0]),
-      row[1],
-      row[2],
-      parseInt(row[3]),
-      parseInt(row[4]),
-      row[5] as MatchResult,
-      row[6]
-    ];
-  }
+export class MatchReader {
+  constructor(public reder: DataReader) {}
 }
